Added HW2 without static sys_print
