+++ syscall.c
@@ -3,6 +3,6 @@
 
 #include <unistd.h>
 
-static int sys_print(char *str, int len) {
+int sys_print(char *str, int len) {
 	return write(1, str, len);
 }
